fix: pubsub child updates
